backend/pyproject.toml
[project]
name = "kobi-agent-backend"
version = "0.1.0"
description = "KOBİ AI Agent — FastAPI Backend"
authors = [{ name = "KOBİ AI Team" }]
requires-python = ">=3.12"
dynamic = ["dependencies"]

[tool.poetry]
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.12"

# Web framework
fastapi = { extras = ["all"], version = "^0.115.0" }
uvicorn = { extras = ["standard"], version = "^0.34.0" }

# ORM & Database
sqlalchemy = { extras = ["asyncio"], version = "^2.0.0" }
asyncpg = "^0.30.0"
alembic = "^1.14.0"

# Validation & Settings
pydantic-settings = "^2.7.0"

# Security
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
passlib = { extras = ["bcrypt"], version = "^1.7.4" }

# Cache
redis = { extras = ["asyncio"], version = "^5.2.0" }

# HTTP Client (dış servis çağrıları için)
httpx = "^0.28.0"

# XSS Sanitization
bleach = "^6.2.0"

# LLM SDK (Anthropic Claude)
anthropic = "^0.42.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
ruff = "^0.8.0"

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
